% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetLM.R
\name{flatten_params}
\alias{flatten_params}
\title{Flattens the network parameters so they can be passed to [minipack.lm::nls.lm]
'par' argument}
\usage{
flatten_params(object)
}
\arguments{
\item{object}{an object of class "nnetLM"}
}
\value{
flattened vector with network parameters (weights and biases)
}
\description{
Flattens the network parameters so they can be passed to [minipack.lm::nls.lm]
'par' argument
}
