% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetLM.R
\name{residFun}
\alias{residFun}
\title{Residual function needed by [minipack.lm::nls.lm]}
\usage{
residFun(params, observed, object, xx)
}
\arguments{
\item{params}{vector of flattened network parameters (weights and biases)}

\item{observed}{an object of class "nnetLM"}

\item{object}{an object of class "nnetLM"}

\item{xx}{an object of class "nnetLM"}
}
\value{
unflattened list with network parameters (weights and biases)
}
\description{
Residual function needed by [minipack.lm::nls.lm]
}
\seealso{
[flatten_params()]
}
