% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetLM.R
\name{train.nnetLM}
\alias{train.nnetLM}
\title{Train the neural network with Levenberg-Marquardt optimization
using [minipack.lm::nls.lm]}
\usage{
train.nnetLM(object, epochs, progress = FALSE)
}
\arguments{
\item{object}{an object of class "nnetLM"}

\item{epochs}{maximum number of iteration}

\item{progress}{flag for printing network progress. Default is FALSE}
}
\value{
the trained network object
}
\description{
Train the neural network with Levenberg-Marquardt optimization
using [minipack.lm::nls.lm]
}
\seealso{
[minipack.lm::nls.lm()]
}
