% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetLM.R
\name{unflatten_params}
\alias{unflatten_params}
\title{Unflattens the network parameters after they have been used by [minipack.lm::nls.lm]}
\usage{
unflatten_params(params, object)
}
\arguments{
\item{params}{vector of flattened network parameters (weights and biases)}

\item{object}{an object of class "nnetLM"}
}
\value{
unflattened list with network parameters (weights and biases)
}
\description{
Unflattens the network parameters after they have been used by [minipack.lm::nls.lm]
}
