% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetLM.R
\name{predict.nnetLM}
\alias{predict.nnetLM}
\title{Performs a forward pass}
\usage{
\method{predict}{nnetLM}(object, newdata, params)
}
\arguments{
\item{object}{an object of class "nnetLM"}

\item{newdata}{Matrix of predictors}

\item{params}{vector of flattened network parameters (weights and biases)}
}
\value{
An object with S3 class "nnetLM"
}
\description{
Performs a forward pass
}
\examples{
set.seed(123)
x <- seq(-10, 10, by = 0.1)
y <- sin(x) + rnorm(length(x), mean = 0, sd = 0.1)
X <- matrix(x, nrow = length(x), ncol = 1)
hidden <- c(10)
actFn <- c("tanh", "linear")
nnet.obj <- nnetLM(X, y, hidden, actFn)
pred.nnetLM <- predict(nnet.obj, X, nnetLM:::flatten_params(nnet.obj))
}
\seealso{
[flatten_params()]
}
